# api/openapi/v1/openapi.yaml
openapi: 3.0.3
info:
  title: DBack Control Plane API
  version: 1.0.0
  description: REST API for DBack server control plane (v1).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /operations:
    get:
      summary: List operations
      responses:
        "200":
          description: Paginated operations
    post:
      summary: Create operation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, profile_id]
              properties:
                kind:
                  type: string
                profile_id:
                  type: string
      responses:
        "202":
          description: Accepted
...
  /operations/{id}:
    get:
      summary: Get operation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Operation
  /operations/{id}/cancel:
    post:
      summary: Cancel operation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Canceled
...